abs NEEDS TO BE fabs everywhere. I fixed a bug regarding this, also made it only check useful reps. Still may not be working for  > 8$, no idea why. It finishes, but perhaps returns an absurd result for $ with shape $. Running it again from scratch, we'll see.

+++ localCode/Vector.hpp
@@ -4,7 +4,7 @@
 #include <vector>
 #include <math.h>
 
-typedef float scalar;
+typedef double scalar;
 typedef std::vector<scalar> vec;
 
 scalar norm(const vec& v);
